MOVE linkdrop acc manager to own file
